core-docker-compose-generator/config-generator/requirements.txt
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
jinja2
